Реализовать работу с полем Views, сохранение и получение статистики просмотров

+++ main-service/main-dto/src/main/java/ru/practicum/dto/event/EventFullDto.java
@@ -42,5 +42,5 @@ public class EventFullDto {
     String state;
     @NotBlank
     String title;
-    int views;
+    Long views;
 }
