+++ main-service/main-dto/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -32,5 +32,5 @@ public class EventShortDto {
     Boolean paid;
     @NotBlank
     String title;
-    int views;
+    Long views;
 }
